requirements.txt
twelvelabs
python-dotenv
google-genai